requirements.txt
openai
pandas
numpy
networkx
scikit-learn
jinja2
python-dotenv
sentence-transformers
